Moved the calls to papaparse inside the webworker
