+++ app/client/package.json
@@ -62,7 +62,6 @@
     "glossary-panel": "github:Eastern-Research-Group/glossary",
     "highcharts": "9.3.3",
     "highcharts-react-official": "3.1.0",
-    "papaparse": "^5.3.2",
     "rc-slider": "^10.0.0",
     "rc-tooltip": "^5.1.1",
     "react": "17.0.2",
